wip: refactor files and types
